backend/summarise/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
celery
python-docx
PyPDF2
huggingface_hub
